+++ src/ts/components/core/modal/Modal.tsx
@@ -3,7 +3,7 @@ import { DashBaseProps } from 'props/dash';
 import { ModalProps } from 'props/modal';
 import { StylesApiProps } from 'props/styles';
 import React, { useEffect, useState } from 'react';
-import { getLoadingState } from '../../utils/dash3';
+import { getLoadingState } from '../../../utils/dash3';
 
 interface Props extends ModalProps, StylesApiProps, DashBaseProps {}